Java: word-count #2, 'refactor' for efficiency reasons
Technically this is better, as we only process the stream of words. This
'reduce' is what might be called a 'fold' in other languages like
Kotlin, Scala or (I'm lead to believe) Haskell. The main thing issue I
have with this is the ugliness of it all. The signatures on the
accumulator and combiner functions are something else. However, in terms
of the mechanics of the operation are concerned, it's what I was aiming
for.